Fix a bug with [WARN] prefix missing whitespace
use of logger.Print causes spaces between args only when the 2 args are
not both strings. Instead, we should use logger.Println which always
causes a space.

While we're at it, ensure our forming of new array is as
efficient as it can be.

Also moves Godeps to top of repo and adds tests w/ circle.